Incorporated MVC pattern
